Address Copilot review comments.
