+++ _frontend/src/components/node/Network_RegisteredNodes.vue
@@ -8,11 +8,11 @@
 
   <DashboardCardV2>
     <template #title>
-      <span>{{ `${props.nodes.length}  ${props.title}` }}</span>
+      <span>{{ `${filteredNodes.length}  ${props.title}` }}</span>
     </template>
     <template #content>
       <p class="mb-4 h-is-low-contrast"> {{ props.subtitle }}</p>
-      <RegisteredNodeTable :filter-service-type="props.filterServiceType" :nodes="props.nodes"/>
+      <RegisteredNodeTable :filter-service-type="props.filterServiceType" :nodes="filteredNodes"/>
     </template>
   </DashboardCardV2>
 
@@ -25,7 +25,7 @@
 <script lang="ts" setup>
 
 import DashboardCardV2 from "@/components/DashboardCardV2.vue";
-import {PropType} from "vue";
+import {computed, PropType} from "vue";
 import RegisteredNodeTable from "@/components/node/RegisteredNodeTable.vue";
 import {RegisteredNode, RegisteredNodeType} from "@/schemas/MirrorNodeSchemas.ts";
 
@@ -40,6 +40,15 @@ const props = defineProps({
     type: Object as PropType<RegisteredNodeType | null>,
     default: null
   }
+})
+
+const filteredNodes = computed(() => {
+  if (props.filterServiceType === null) {
+    return props.nodes
+  }
+  return props.nodes.filter(node => {
+    return node.service_endpoints.some(endPoint => endPoint.type === props.filterServiceType)
+  })
 })
 
 </script>
